#117 PosterRequest -> PosterRequest, PosterRequestImpl
